+++ src/app/page.tsx
@@ -1,7 +1,9 @@
+import * as React from 'react';
+
 export default function RootPage() {
   return (
     <main className="grid h-screen place-content-center">
-      <h1 className="text-center font-display">Rahayu by Devsantara</h1>
+      <h1 className="font-display text-center">Rahayu by Devsantara</h1>
     </main>
   );
 }
