+++ tsconfig.json
@@ -15,7 +15,7 @@
     "jsx": "preserve",
     "plugins": [{ "name": "next" }],
     "paths": {
-      "~/*": ["./*"]
+      "~/*": ["./src/*"]
     }
   },
   "include": ["next-env.d.ts", ".next/types/**/*.ts", "**/*.ts", "**/*.tsx"],
